backend/requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
fastapi==0.61.1
h11==0.11.0
idna==2.10
loguru==0.5.3
passlib==1.7.4
psycopg2-binary==2.8.6
pydantic==1.6.1
PyJWT==1.7.1
python-multipart==0.0.5
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.19
starlette==0.13.6
typing-extensions==3.7.4.3
urllib3==1.25.10
uvicorn==0.12.1
